new minishell

+++ libft/srcs/ft_print/ft_putstr.c
@@ -6,7 +6,7 @@
 /*   By: lvirgini <lvirgini@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/18 14:16:34 by lvirgini          #+#    #+#             */
-/*   Updated: 2020/03/07 13:07:09 by lvirgini         ###   ########.fr       */
+/*   Updated: 2021/06/21 18:17:05 by lvirgini         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 
 void	ft_putstr(char *s)
 {
-	unsigned int i;
+	size_t	i;
 
 	if (s)
 	{
