+++ libft/srcs/ft_string/ft_strchr_len.c
@@ -6,18 +6,20 @@
 /*   By: lvirgini <lvirgini@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/07 14:01:05 by lvirgini          #+#    #+#             */
-/*   Updated: 2020/03/07 14:02:04 by lvirgini         ###   ########.fr       */
+/*   Updated: 2021/06/21 18:33:17 by lvirgini         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "libft.h"
 
 /*
 ** 	Recherche un caractère (c) dans une chaine de caratere (*s).
 ** 	Renvoie l'index de la première occurence, sinon -1;
 */
 
-int			ft_strchr_len(const char *s, int c)
+int	ft_strchr_len(const char *s, int c)
 {
-	unsigned int i;
+	size_t	i;
 
 	i = 0;
 	while (s && s[i])
